+++ R/grid_search.R
@@ -1,3 +1,24 @@
+## Particle filter outputs
+#' @noRd
+beta_date_particle_filter <- function(beta, start_date, 
+                                      model_params, data, 
+                                      pars_obs, n_particles) {
+  
+  # Edit beta in parameters
+  model_params$beta_y <- beta
+  model_params$beta_t <- 0
+  
+  # Objects used by particle filter run
+  pf_data <- particle_filter_data(data, start_date, steps_per_day= 1/model_params$dt)
+  model_run <- sircovid(params = model_params)
+  compare_func <- compare_icu(model_run, pars_obs, pf_data)
+  
+  X <- particle_filter(pf_data, model_run, compare_func, 
+                       n_particles = n_particles, save_particles = TRUE)
+  
+}
+
+
 ##' Run a grid search of the particle filter over beta and start date.
 ##' This is parallelised, first run \code{plan(multiprocess)} to set 
 ##' this up.
@@ -26,6 +47,8 @@
 ##' @param pars_obs list of parameters to use in comparison
 ##'   with \code{compare_icu}.
 ##'   
+##' @param n_particles Number of particles. Positive Integer. Default = 100
+##' 
 ##' @return List of beta and start date grid values, and
 ##'   normalised probabilities at each point
 ##' 
@@ -40,7 +63,8 @@ scan_beta_date <- function(
   day_step,
   data, 
   model_params = NULL,
-  pars_obs = NULL) {
+  pars_obs = NULL,
+  n_particles = 100) {
   #
   # Set up parameter space to scan
   #
@@ -84,20 +108,11 @@ scan_beta_date <- function(
   #
   ## Particle filter outputs
   pf_run_outputs <- furrr::future_pmap(
-    param_grid,
-    function(beta, start_date) {
-      # Edit beta in parameters
-      model_params$beta_y <- beta
-      model_params$beta_t <- 0
-      
-      # Objects used by particle filter run
-      pf_data <- particle_filter_data(data, start_date, steps_per_day= 1/model_params$dt)
-      model_run <- sircovid(params = model_params)
-      compare_func <- compare_icu(model_run, pars_obs, pf_data)
-      
-      X <- particle_filter(pf_data, model_run, compare_func, n_particles=100)
-    }
+    .l = param_grid, .f = beta_date_particle_filter,
+    model_params = model_params, data = data, 
+    pars_obs = pars_obs, n_particles = n_particles
   )
+  
   ## Extract Log-likelihoods
   log_ll <- furrr::future_map_dbl(pf_run_outputs, ~ .$log_likelihood) 
   
@@ -114,10 +129,11 @@ scan_beta_date <- function(
   prob_matrix <- exp(mat_log_ll)
   renorm_mat_LL <- prob_matrix/sum(prob_matrix)
   
-  results <- list(x=beta_1D, 
-                  y=date_list,
-                  mat_log_ll=mat_log_ll,
-                  renorm_mat_LL=renorm_mat_LL,
+
+  results <- list(x = beta_1D, 
+                  y = date_list,
+                  mat_log_ll = mat_log_ll,
+                  renorm_mat_LL = renorm_mat_LL,
                   inputs = list(
                     model_params = model_params,
                     pars_obs = pars_obs,
@@ -138,3 +154,35 @@ plot.sircovid_scan <- function(x, ..., what = "likelihood") {
 }
 
 
+##' @export
+plot.sample_grid_search <- function(x, ..., what = "ICU") {
+  
+  # what are we plotting
+  
+  if (what == "ICU") {
+    
+    index <- c(odin_index(x$inputs$model)$I_ICU) - 1L
+    ylab <- "ICU"
+    particles <- vapply(seq_len(dim(x$trajectories)[3]), function(y) {
+      rowSums(x$trajectories[,index,y], na.rm = TRUE)}, 
+      FUN.VALUE = numeric(dim(x$trajectories)[1]))
+    plot_particles(particles, ylab = ylab)
+    points(as.Date(x$inputs$data$date), x$inputs$data$itu / x$inputs$pars_obs$phi_ICU, pch = 19)
+    
+  } else if(what == "Deaths") {
+    
+    index <- c(odin_index(x$inputs$model)$D) - 1L
+    ylab <- "Deaths"
+    particles <- vapply(seq_len(dim(x$trajectories)[3]), function(y) {
+      rowSums(x$trajectories[,index,y], na.rm = TRUE)}, 
+      FUN.VALUE = numeric(dim(x$trajectories)[1]))
+    plot_particles(particles, ylab = ylab)
+    points(as.Date(x$inputs$data$date), cumsum(x$inputs$data$deaths), pch = 19)
+    
+  } else {
+    
+    stop("Requested what must be one of 'ICU' or 'Deaths'")
+    
+  } 
+  
+}